+++ Source/Boilerplate.Templates/Content/ApiTemplate/Startup.cs
@@ -209,7 +209,7 @@
                 .UseMvc()
                 // Add Swagger to the request pipeline.
                 .UseSwagger()
-                .UseSwaggerUi(
+                .UseSwaggerUI(
                     options =>
                     {
                         options.SwaggerEndpoint("/swagger/v1/swagger.json", "Version 1");