+++ RELEASE NOTES.md
@@ -1,6 +1,7 @@
 # Boilerplate.Templates 1.4.0
 - Added OPTIONS endpoint.
 - Added HEAD endpoints to all GET endpoints.
+- Update Swashbuckle.AspNetCore to 1.0.0-rc3
 
 # Boilerplate.Templates 1.3.0
 - Added Versioning feature.